Loop bug

+++ src/libft/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: pvaladar <pvaladar@student.42lisboa.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/16 16:40:15 by pvaladar          #+#    #+#             */
-/*   Updated: 2022/03/10 16:10:19 by pvaladar         ###   ########.fr       */
+/*   Updated: 2022/06/28 14:42:30 by pvaladar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -39,7 +39,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
